% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{bwd_stepwise}
\alias{bwd_stepwise}
\title{Backward stepwise selection}
\usage{
bwd_stepwise(fwd, control)
}
\arguments{
\item{fwd}{output from forward selection}

\item{control}{mars control object}
}
\value{
list of y, B, Bfuncs outputs from backward selection
}
\description{
Backward stepwise selection
}
