% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{fwd_stepwise}
\alias{fwd_stepwise}
\title{Forward stepwise selection}
\usage{
fwd_stepwise(y, x, control = mars.control())
}
\arguments{
\item{y}{response data}

\item{x}{explanatory data}

\item{control}{mars control object created with EHC.MARS::mars.control function}
}
\value{
list of y,B, and Bfuncs objects from forward selection for mars
}
\description{
Forward stepwise selection
}
