% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{init_B}
\alias{init_B}
\title{init_B}
\usage{
init_B(N, Mmax)
}
\arguments{
\item{N}{number of rows for data matrix}

\item{Mmax}{number of columns not including intercept}
}
\value{
initialized matrix
}
\description{
init_B
}
