% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{LOF}
\alias{LOF}
\title{LOF}
\usage{
LOF(formula, data, control)
}
\arguments{
\item{formula}{formula used in linear model}

\item{data}{dataset}

\item{control}{mars control object, created with EHC.MARS::mars.control function}
}
\value{
GCV criterion value
}
\description{
LOF
}
