% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{make_B}
\alias{make_B}
\title{make_B}
\usage{
make_B(X, Bfuncs)
}
\arguments{
\item{X}{data matrix from mars}

\item{Bfuncs}{Bfuncs data from mars object}
}
\value{
Initialized matrix for predict.mars function
}
\description{
Function used in predict.mars to create initial matrix
}
