% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{Mars}
\usage{
mars(formula, data, control = mars.control())
}
\arguments{
\item{formula}{a formula}

\item{data}{dataset}

\item{control}{mars control object (can be created using mars.control)}
}
\value{
Mars returns ...
}
\description{
Description of what mars is
}
