% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{new_mars.control}
\alias{new_mars.control}
\title{Mars Constructor}
\usage{
new_mars.control(control)
}
\arguments{
\item{control}{a mars control object created with EHC.MARS::mars.control function}
}
\value{
mars control object
}
\description{
Mars Constructor
}
