% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mars.R
\name{plot.mars}
\alias{plot.mars}
\title{plot.mars}
\usage{
\method{plot}{mars}(mars_obj)
}
\arguments{
\item{mars_obj}{A mars class object, created using the EHC.MARS::mars() function}
}
\value{
Gives set of plots for mars object function
}
\description{
plot.mars
}
\examples{
mars_object<-mars(formula, data, control = mars.control()) 
plot(mars_object)
}
