% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mars.R
\name{print.mars}
\alias{print.mars}
\title{print.mars}
\usage{
\method{print}{mars}(mars_obj)
}
\arguments{
\item{mars_obj}{A mars class object, created using the EHC.MARS::mars() function}
}
\value{
Outputs the call and coefficients used for creating mars object
}
\description{
print.mars
}
\examples{
mars_object<-mars(formula, data, control = mars.control()) 
print(mars_object)
}
