% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mars.R
\name{summary.mars}
\alias{summary.mars}
\title{Summary.mars}
\usage{
\method{summary}{mars}(mars_obj)
}
\arguments{
\item{mars_obj}{A mars class object, created using the EHC.MARS::mars() function}
}
\value{
Outputs a summary of the mars object
}
\description{
Summary.mars
}
\examples{
mars_object<-mars(formula, data, control = mars.control()) 
summary(mars_object)
}
