% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{validate_mars.control}
\alias{validate_mars.control}
\title{Mars Validator}
\usage{
validate_mars.control(control)
}
\arguments{
\item{control}{Mars control object created with EHC.MARS::mars.control function}
}
\value{
mars control object
}
\description{
Mars Validator
}
